Move truffle artifact from util to module
